refactoring the catalogs class
