+++ astroquery/mast/discovery_portal.py
@@ -21,6 +21,7 @@ from astropy.utils import deprecated
 from ..query import BaseQuery, QueryWithLogin
 from ..utils import async_to_sync
 from ..utils.class_or_instance import class_or_instance
+from ..exceptions import InputWarning, NoResultsWarning
 
 from . import conf, utils
